Allowing multiple UUID versions to validate, including the NIL value (#145)
* Allowing multiple UUID versions to validate, including the NIL value.

This is done by changing the constant to a bitmask which may be combined. By combining multiple masks using the | operator you can attempt multiple UUID versions.

The most common usecase will probably be to check any of the versions with NULL. For example: Uuid::UUID_NIL | Uuid::UUID_V4

This does limit error reporting a bit. As the check for correct version-number gives less information and the error-message given can return multiple version numbers now.

* Not sure how this newline got there.

* Add invalid format tests and extend the version exception message

* Fixed mistake in version validation, should not be less then 0.

* Added documentation.

* Added the more permissive UUID_VALID regex and made it the default.

As discussed with @RickvdStaaij

* Test valid format, make valid format default, fix error bug, update docs

* Fix broken tests

* Fix casing on Uuid class in the documentation